requirements.txt
# Core RAG stack
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.0

# Vector store
chromadb>=0.5.0

# Embeddings
sentence-transformers>=3.0.0
huggingface-hub>=0.24.0
google-generativeai>=0.7.2
langchain-google-genai>=0.0.11

# PDF loader dependency
pypdf>=4.0.0

# App & utils
streamlit>=1.37.0
python-dotenv>=1.0.1
